+++ src/Text/Show/Text/Data/List/NonEmpty.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE CPP, TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-|
 Module:      Text.Show.Text.Data.List.NonEmpty
@@ -18,8 +18,10 @@ import Data.List.NonEmpty (NonEmpty)
 
 import Prelude hiding (Show)
 
-import Text.Show.Text (Show(showbPrec), Builder)
+import Text.Show.Text (Show(showbPrec), Show1(showbPrec1), Builder)
 import Text.Show.Text.TH (deriveShowPragmas, defaultInlineShowbPrec)
+
+#include "inline.h"
 
 -- | Convert a 'NonEmpty' list to a 'Builder' with the given precedence.
 -- 
@@ -29,3 +31,7 @@ showbNonEmptyPrec = showbPrec
 {-# INLINE showbNonEmptyPrec #-}
 
 $(deriveShowPragmas defaultInlineShowbPrec ''NonEmpty)
+
+instance Show1 NonEmpty where
+    showbPrec1 = showbPrec
+    INLINE_INST_FUN(showbPrec1)