+++ tests/Instances/Data/Semigroup.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE GeneralizedNewtypeDeriving, StandaloneDeriving #-}
+{-# LANGUAGE CPP, GeneralizedNewtypeDeriving, StandaloneDeriving #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-|
 Module:      Instances.Data.Semigroup
@@ -12,9 +12,15 @@ Provides 'Arbitrary' instances for @Semigroup@ data types.
 -}
 module Instances.Data.Semigroup () where
 
-import Data.Semigroup (Min(..), Max(..), First(..),
-                       Last(..), WrappedMonoid(..), Option(..))
-import Test.Tasty.QuickCheck (Arbitrary)
+#if !(MIN_VERSION_base(4,8,0))
+import Control.Applicative ((<*>))
+#endif
+
+import Data.Functor ((<$>))
+import Data.Semigroup (Min(..), Max(..), First(..), Last(..),
+                       WrappedMonoid(..), Option(..), Arg(..))
+
+import Test.Tasty.QuickCheck (Arbitrary(..))
 
 deriving instance Arbitrary a => Arbitrary (Min a)
 deriving instance Arbitrary a => Arbitrary (Max a)
@@ -22,3 +28,6 @@ deriving instance Arbitrary a => Arbitrary (First a)
 deriving instance Arbitrary a => Arbitrary (Last a)
 deriving instance Arbitrary m => Arbitrary (WrappedMonoid m)
 deriving instance Arbitrary a => Arbitrary (Option a)
+
+instance (Arbitrary a, Arbitrary b) => Arbitrary (Arg a b) where
+    arbitrary = Arg <$> arbitrary <*> arbitrary